+++ __test__/alph_case.test.js
@@ -1,0 +1,6 @@
+const alphCase = require('../app/alph_case');
+const alphabet = 'abcdefghijklmnopqrstuvwxyz';
+
+test('alphabet case', () => {
+  expect(alphCase('A', alphabet)).toBe(alphabet.toUpperCase());
+});
