+++ app/custom_transform_stream.js
@@ -1,8 +1,6 @@
 const { Transform } = require('stream');
-const atbash = require('./atbash');
-const encode = require('./encode');
-const decode = require('./decode');
-const alphCase = require('./alph_case');
+
+const transformLetter = require('./transform_letter');
 
 class CustomTransform extends Transform {
   constructor(cipher) {
@@ -16,26 +14,6 @@ class CustomTransform extends Transform {
       .map((letter) => transformLetter(letter, this.cipher))
       .join('');
     callback(null, data);
-  }
-}
-
-function transformLetter(letter, cipher) {
-  const alphabet = 'abcdefghijklmnopqrstuvwxyz';
-  const alphLen = 26;
-  if (!alphabet.toLowerCase().includes(letter.toLowerCase())) return letter;
-  const letterIndex = alphabet.indexOf(letter.toLowerCase());
-  const newAlphabet = alphCase(letter, alphabet);
-  switch (cipher) {
-    case 'A':
-      return atbash(letterIndex, newAlphabet);
-    case 'C1':
-      return encode(1, newAlphabet, alphLen, letterIndex);
-    case 'R1':
-      return encode(8, newAlphabet, alphLen, letterIndex);
-    case 'C0':
-      return decode(1, newAlphabet, letterIndex);
-    case 'R0':
-      return decode(8, newAlphabet, letterIndex);
   }
 }
